[Kernel] Move kernel test helpers to sglang.test.kernels (out of the runtime kernels package)
The shared test-support helpers do not belong under the runtime `kernels/`
package. Move python/sglang/kernels/testing/ -> python/sglang/test/kernels/
(utils, kv_canary/, deepseek_v4/), the established in-package test-support
namespace (alongside test_utils, ci_register, runners, mock_model, ...), and
rewrite sglang.kernels.testing -> sglang.test.kernels imports.

Kept separate from the existing srt-level sglang/test/kv_canary/ (e2e fixtures):
these are the kernel-level fuzz/oracle helpers, so they live under
sglang/test/kernels/kv_canary/ mirroring the kernel test tree
test/registered/kernels/. Still an importable installed package (required:
test/registered/ tests import it by absolute path).
