+++ python/sglang/test/kernels/kv_canary/_fixtures.py
@@ -10,7 +10,7 @@ from sglang.kernels.ops.kv_canary.verify import (
     VerifyPlan,
 )
 from sglang.kernels.ops.kv_canary.write import WritePlan
-from sglang.kernels.testing.kv_canary._constants import DEFAULT_NUM_SLOTS
+from sglang.test.kernels.kv_canary._constants import DEFAULT_NUM_SLOTS
 
 _DEVICE = torch.device("cuda")
