+++ python/sglang/test/kernels/kv_canary/_hand_oracle.py
