+++ test/registered/kernels/ops/attention/test_c4_v2.py
@@ -8,7 +8,9 @@ import torch
 import triton
 
 from sglang.kernels.ops.attention.dsv4 import compress_forward
-from sglang.kernels.testing.deepseek_v4.common import (
+from sglang.srt.utils import get_device
+from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
+from sglang.test.kernels.deepseek_v4.common import (
     LegacyContext,
     PagedContext,
     make_legacy_context,
@@ -16,8 +18,6 @@ from sglang.kernels.testing.deepseek_v4.common import (
     make_state_pool,
     to_seq_extend,
 )
-from sglang.srt.utils import get_device
-from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
 
 register_cuda_ci(est_time=30, stage="base-b-kernel-unit", runner_config="1-gpu-large")
 register_amd_ci(est_time=30, suite="nightly-amd-kernel-1-gpu", nightly=True)
