+++ test/registered/kernels/ops/communication/test_tp_qknorm.py
@@ -21,11 +21,11 @@ from sglang.kernels.ops.communication.all_reduce import (
     get_fused_parallel_qknorm_module,
 )
 from sglang.kernels.ops.communication.mp import register_comm_cleanup
-from sglang.kernels.testing.utils import multigpu_pytest_main
 from sglang.srt.distributed.device_communicators.custom_all_reduce_v2 import (
     CustomAllReduceV2,
 )
 from sglang.test.ci.ci_register import register_cuda_ci
+from sglang.test.kernels.utils import multigpu_pytest_main
 
 register_cuda_ci(
     est_time=300,
