+++ test/registered/kernels/ops/kv_canary/test_kernel_config.py
@@ -11,7 +11,8 @@ from sglang.kernels.ops.kv_canary.verify import (
     launch_canary_verify_kernel,
 )
 from sglang.kernels.ops.kv_canary.write import WritePlan
-from sglang.kernels.testing.kv_canary._canary_helpers import (
+from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
+from sglang.test.kernels.kv_canary._canary_helpers import (
     FakeViolationLog,
     assert_canary_buf_equal,
     assert_canary_state_equal,
@@ -23,17 +24,16 @@ from sglang.kernels.testing.kv_canary._canary_helpers import (
     make_write_plan_pair,
     stamp_clean_chain,
 )
-from sglang.kernels.testing.kv_canary._differential import (
+from sglang.test.kernels.kv_canary._differential import (
     _assert_plans_byte_equal,
     _run_both_plan,
     _run_both_verify,
     _run_both_write,
 )
-from sglang.kernels.testing.kv_canary._fixtures import (
+from sglang.test.kernels.kv_canary._fixtures import (
     dummy_pseudo_tensors,
     empty_extras,
 )
-from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
 
 register_cuda_ci(est_time=60, stage="base-b-kernel-unit", runner_config="1-gpu-large")
 register_amd_ci(est_time=60, stage="jit-kernel-unit", runner_config="amd")
