+++ test/registered/kernels/ops/kv_canary/test_plan_fuzz.py
@@ -7,20 +7,20 @@ from typing import Optional
 import pytest
 import torch
 
-from sglang.kernels.testing.kv_canary._differential import _run_both_plan
-from sglang.kernels.testing.kv_canary._fixtures import (
+from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
+from sglang.test.kernels.kv_canary._differential import _run_both_plan
+from sglang.test.kernels.kv_canary._fixtures import (
     allocate_plan_pair,
     derive_plan_capacity,
     make_lut,
     make_padding_mask,
     make_req_to_token,
 )
-from sglang.kernels.testing.kv_canary._fuzz_driver import (
+from sglang.test.kernels.kv_canary._fuzz_driver import (
     FUZZ_SEEDS_PR,
     run_fuzz_combo,
 )
-from sglang.kernels.testing.kv_canary._invariants import PlanInvariants
-from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
+from sglang.test.kernels.kv_canary._invariants import PlanInvariants
 
 register_cuda_ci(est_time=30, stage="base-b-kernel-unit", runner_config="1-gpu-large")
 register_amd_ci(est_time=30, stage="jit-kernel-unit", runner_config="amd")
