+++ test/registered/kernels/ops/kv_canary/test_write_hand.py
@@ -20,7 +20,8 @@ from sglang.kernels.ops.kv_canary.verify import (
 from sglang.kernels.ops.kv_canary.write import (
     launch_canary_write_kernel,
 )
-from sglang.kernels.testing.kv_canary._canary_helpers import (
+from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
+from sglang.test.kernels.kv_canary._canary_helpers import (
     FakeViolationLog,
     assert_canary_state_equal,
     assert_only_bits_set,
@@ -37,19 +38,18 @@ from sglang.kernels.testing.kv_canary._canary_helpers import (
     stamp_pair,
     to_signed_int64,
 )
-from sglang.kernels.testing.kv_canary._differential import (
+from sglang.test.kernels.kv_canary._differential import (
     _run_both_write,
     run_write_diff,
 )
-from sglang.kernels.testing.kv_canary._fixtures import (
+from sglang.test.kernels.kv_canary._fixtures import (
     clone_real_kv_sources,
     dummy_pseudo_tensors,
 )
-from sglang.kernels.testing.kv_canary._hand_oracle import (
+from sglang.test.kernels.kv_canary._hand_oracle import (
     _hand_fold_all,
     _hand_fold_partial,
 )
-from sglang.test.ci.ci_register import register_amd_ci, register_cuda_ci
 
 register_cuda_ci(est_time=30, stage="base-b-kernel-unit", runner_config="1-gpu-large")
 register_amd_ci(est_time=30, stage="jit-kernel-unit", runner_config="amd")
