+++ test/registered/unit/distributed/test_vmm_utils.py
@@ -21,7 +21,6 @@ import torch.distributed as dist
 from cuda.bindings import driver as drv
 
 from sglang.kernels.jit.utils import cache_once
-from sglang.kernels.testing.utils import multigpu_pytest_main
 from sglang.srt.distributed.device_communicators.vmm_utils import (
     check_drv,
     exchange_posix_fds,
@@ -32,6 +31,7 @@ from sglang.srt.distributed.device_communicators.vmm_utils import (
     release_mappings,
 )
 from sglang.test.ci.ci_register import register_cuda_ci
+from sglang.test.kernels.utils import multigpu_pytest_main
 
 register_cuda_ci(est_time=60, stage="base-b", runner_config="2-gpu-large")
